+++ ecmascript/transforms/optimization/tests/simplify.rs
@@ -510,3 +510,36 @@ test!(
     new A();
     "
 );
+
+test!(
+    syntax(),
+    |_| chain!(
+        strip(),
+        decorators(Default::default()),
+        class_properties(),
+        simplifier(Default::default()),
+        compat::es2018(),
+        compat::es2017(),
+        compat::es2016(),
+        compat::es2015(Mark::fresh(Mark::root()), Default::default()),
+        compat::es3(true),
+        import_analyzer(),
+        inject_helpers(),
+        common_js(Mark::fresh(Mark::root()), Default::default()),
+    ),
+    issue_389_3,
+    "
+import Foo from 'foo';
+Foo.bar = true;
+",
+    "
+'use strict';
+var _foo = _interopRequireDefault(require('foo'));
+function _interopRequireDefault(obj) {
+    return obj && obj.__esModule ? obj : {
+        default: obj
+    };
+}
+_foo.default.bar = true;
+"
+);
