+++ ecmascript/transforms/typescript/tests/strip.rs
@@ -1,11 +1,13 @@
 use swc_common::chain;
 use swc_ecma_parser::{Syntax, TsConfig};
 use swc_ecma_transforms_base::resolver::resolver;
+use swc_ecma_transforms_compat::es2020::class_properties;
 use swc_ecma_transforms_compat::es2020::typescript_class_properties;
 use swc_ecma_transforms_proposal::decorators;
 use swc_ecma_transforms_testing::test;
 use swc_ecma_transforms_testing::test_exec;
 use swc_ecma_transforms_typescript::strip;
+use swc_ecma_transforms_typescript::strip::strip_with_config;
 use swc_ecma_visit::Fold;
 
 fn tr() -> impl Fold {
@@ -3200,37 +3202,4 @@ test!(
     });
     console.log(submissions);
     "#
-);
-
-test!(
-    syntax(),
-    |_| chain!(
-        typescript::strip(),
-        decorators(Default::default()),
-        class_properties(),
-        simplifier(Default::default()),
-        compat::es2018(),
-        compat::es2017(),
-        compat::es2016(),
-        compat::es2015(Mark::fresh(Mark::root()), Default::default()),
-        compat::es3(true),
-        import_analyzer(),
-        inject_helpers(),
-        common_js(Mark::fresh(Mark::root()), Default::default()),
-    ),
-    issue_389_3,
-    "
-import Foo from 'foo';
-Foo.bar = true;
-",
-    "
-'use strict';
-var _foo = _interopRequireDefault(require('foo'));
-function _interopRequireDefault(obj) {
-    return obj && obj.__esModule ? obj : {
-        default: obj
-    };
-}
-_foo.default.bar = true;
-"
 );
